+++ packages/carp_audio_package/lib/audio.g.dart
@@ -74,6 +74,54 @@ Map<String, dynamic> _$NoiseDatumToJson(NoiseDatum instance) {
   return val;
 }
 
+VideoDatum _$VideoDatumFromJson(Map<String, dynamic> json) => VideoDatum(
+      filename: json['filename'] as String,
+      videoType: $enumDecode(_$VideoTypeEnumMap, json['video_type']),
+      startRecordingTime: json['start_recording_time'] == null
+          ? null
+          : DateTime.parse(json['start_recording_time'] as String),
+      endRecordingTime: json['end_recording_time'] == null
+          ? null
+          : DateTime.parse(json['end_recording_time'] as String),
+    )
+      ..id = json['id'] as String?
+      ..timestamp = json['timestamp'] == null
+          ? null
+          : DateTime.parse(json['timestamp'] as String)
+      ..upload = json['upload'] as bool
+      ..metadata = (json['metadata'] as Map<String, dynamic>?)?.map(
+        (k, e) => MapEntry(k, e as String),
+      )
+      ..format = DataFormat.fromJson(json['format'] as Map<String, dynamic>);
+
+Map<String, dynamic> _$VideoDatumToJson(VideoDatum instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
+  val['filename'] = instance.filename;
+  val['upload'] = instance.upload;
+  writeNotNull('metadata', instance.metadata);
+  writeNotNull(
+      'start_recording_time', instance.startRecordingTime?.toIso8601String());
+  writeNotNull(
+      'end_recording_time', instance.endRecordingTime?.toIso8601String());
+  val['format'] = instance.format;
+  val['video_type'] = _$VideoTypeEnumMap[instance.videoType];
+  return val;
+}
+
+const _$VideoTypeEnumMap = {
+  VideoType.video: 'video',
+  VideoType.image: 'image',
+};
+
 NoiseMeasure _$NoiseMeasureFromJson(Map<String, dynamic> json) => NoiseMeasure(
       type: json['type'] as String,
       name: json['name'] as String?,