+++ packages/carp_audio_package/lib/audio_probe.dart
@@ -22,7 +22,6 @@ part of audio;
 ///
 /// The Audio probe generates an [AudioDatum] that holds the meta-data for each
 /// recording along with the actual recording in an audio file.
-/// Audio is recording in an MP3 format.
 /// How to upload this data to a data backend is up to the implementation of the
 /// [DataManager], which is used in the [Study].
 class AudioProbe extends DatumProbe {